Renames encode/decode to transform/restore
