+++ lib/index.js
@@ -418,14 +418,14 @@ Socket.prototype.unsubscribe = function(filter) {
 
 Socket.prototype.send = function(msg, flags) {
   if (Array.isArray(msg)) {
-    msg = msg.map(this.encode);
+    msg = msg.map(this.transform);
     for (var i = 0, len = msg.length; i < len; i++) {
       this.send(msg[i], i < len - 1
         ? zmq.ZMQ_SNDMORE
         : flags);
     }
   } else {
-    msg = this.encode(msg, 0, [msg]);
+    msg = this.transform(msg, 0, [msg]);
     if (!Buffer.isBuffer(msg)) msg = new Buffer(String(msg), 'utf8');
     this._outgoing.push([msg, flags || 0]);
     if (!(flags & zmq.ZMQ_SNDMORE)) this._flush();
@@ -463,7 +463,7 @@ Socket.prototype._flush = function() {
           do {
             emitArgs.push(this._zmq.recv());
           } while (this._receiveMore);
-          emitArgs = ['message'].concat(emitArgs.map(this.decode))
+          emitArgs = ['message'].concat(emitArgs.map(this.restore))
 
           // Handle received message immediately for receive only sockets to prevent memory leak in driver
           if (types[this.type] === zmq.ZMQ_SUB || types[this.type] === zmq.ZMQ_PULL) {
@@ -519,7 +519,7 @@ Socket.prototype.close = function() {
  * @api public
  */
 
-Socket.prototype.encode = function (buf, idx, arr) {
+Socket.prototype.transform = function (buf, idx, arr) {
   return buf;
 };
 
@@ -530,7 +530,7 @@ Socket.prototype.encode = function (buf, idx, arr) {
  * @api public
  */
 
-Socket.prototype.decode = function (buf, idx, arr) {
+Socket.prototype.restore = function (buf, idx, arr) {
   return buf;
 };
